+++ app/src/main/java/com/prey/barcodereader/ui/camera/CameraSource.java
@@ -1,10 +1,19 @@
-/*******************************************************************************
- * Created by Orlando Aliaga
- * Copyright 2016 Prey Inc. All rights reserved.
- * License: GPLv3
- * Full license at "/LICENSE"
- ******************************************************************************/
-package com.prey.barcode.ui.camera;
+/*
+ * Copyright (C) The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.prey.barcodereader.ui.camera;
 
 import android.Manifest;
 import android.annotation.SuppressLint;
@@ -28,7 +37,6 @@ import android.view.WindowManager;
 import com.google.android.gms.common.images.Size;
 import com.google.android.gms.vision.Detector;
 import com.google.android.gms.vision.Frame;
-import com.prey.PreyLogger;
 
 import java.io.IOException;
 import java.lang.Thread.State;
@@ -40,35 +48,67 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
+// Note: This requires Google Play Services 8.1 or higher, due to using indirect byte buffers for
+// storing images.
 
+/**
+ * Manages the camera in conjunction with an underlying
+ * {@link com.google.android.gms.vision.Detector}.  This receives preview frames from the camera at
+ * a specified rate, sending those frames to the detector as fast as it is able to process those
+ * frames.
+ * <p/>
+ * This camera source makes a best effort to manage processing on preview frames as fast as
+ * possible, while at the same time minimizing lag.  As such, frames may be dropped if the detector
+ * is unable to keep up with the rate of frames generated by the camera.  You should use
+ * {@link CameraSource.Builder#setRequestedFps(float)} to specify a frame rate that works well with
+ * the capabilities of the camera hardware and the detector options that you have selected.  If CPU
+ * utilization is higher than you'd like, then you may want to consider reducing FPS.  If the camera
+ * preview or detector results are too "jerky", then you may want to consider increasing FPS.
+ * <p/>
+ * The following Android permission is required to use the camera:
+ * <ul>
+ * <li>android.permissions.CAMERA</li>
+ * </ul>
+ */
+@SuppressWarnings("deprecation")
 public class CameraSource {
     @SuppressLint("InlinedApi")
     public static final int CAMERA_FACING_BACK = CameraInfo.CAMERA_FACING_BACK;
     @SuppressLint("InlinedApi")
     public static final int CAMERA_FACING_FRONT = CameraInfo.CAMERA_FACING_FRONT;
 
+    private static final String TAG = "PREY";
+
+    /**
+     * The dummy surface texture must be assigned a chosen name.  Since we never use an OpenGL
+     * context, we can choose any ID we want here.
+     */
     private static final int DUMMY_TEXTURE_NAME = 100;
 
+    /**
+     * If the absolute difference between a preview size aspect ratio and a picture size aspect
+     * ratio is less than this tolerance, they are considered to be the same aspect ratio.
+     */
     private static final float ASPECT_RATIO_TOLERANCE = 0.01f;
 
     @StringDef({
-            Camera.Parameters.FOCUS_MODE_CONTINUOUS_PICTURE,
-            Camera.Parameters.FOCUS_MODE_CONTINUOUS_VIDEO,
-            Camera.Parameters.FOCUS_MODE_AUTO,
-            Camera.Parameters.FOCUS_MODE_EDOF,
-            Camera.Parameters.FOCUS_MODE_FIXED,
-            Camera.Parameters.FOCUS_MODE_INFINITY,
-            Camera.Parameters.FOCUS_MODE_MACRO
+        Camera.Parameters.FOCUS_MODE_CONTINUOUS_PICTURE,
+        Camera.Parameters.FOCUS_MODE_CONTINUOUS_VIDEO,
+        Camera.Parameters.FOCUS_MODE_AUTO,
+        Camera.Parameters.FOCUS_MODE_EDOF,
+        Camera.Parameters.FOCUS_MODE_FIXED,
+        Camera.Parameters.FOCUS_MODE_INFINITY,
+        Camera.Parameters.FOCUS_MODE_MACRO
     })
     @Retention(RetentionPolicy.SOURCE)
     private @interface FocusMode {}
 
     @StringDef({
-            Camera.Parameters.FLASH_MODE_ON,
-            Camera.Parameters.FLASH_MODE_OFF,
-            Camera.Parameters.FLASH_MODE_AUTO,
-            Camera.Parameters.FLASH_MODE_RED_EYE,
-            Camera.Parameters.FLASH_MODE_TORCH
+        Camera.Parameters.FLASH_MODE_ON,
+        Camera.Parameters.FLASH_MODE_OFF,
+        Camera.Parameters.FLASH_MODE_AUTO,
+        Camera.Parameters.FLASH_MODE_RED_EYE,
+        Camera.Parameters.FLASH_MODE_TORCH
     })
     @Retention(RetentionPolicy.SOURCE)
     private @interface FlashMode {}
@@ -77,14 +117,21 @@ public class CameraSource {
 
     private final Object mCameraLock = new Object();
 
+    // Guarded by mCameraLock
     private Camera mCamera;
 
     private int mFacing = CAMERA_FACING_BACK;
 
+    /**
+     * Rotation of the device, and thus the associated preview images captured from the device.
+     * See {@link Frame.Metadata#getRotation()}.
+     */
     private int mRotation;
 
     private Size mPreviewSize;
 
+    // These values may be requested by the caller.  Due to hardware limitations, we may need to
+    // select close, but not exactly the same values for these.
     private float mRequestedFps = 30.0f;
     private int mRequestedPreviewWidth = 1024;
     private int mRequestedPreviewHeight = 768;
@@ -93,19 +140,41 @@ public class CameraSource {
     private String mFocusMode = null;
     private String mFlashMode = null;
 
+    // These instances need to be held onto to avoid GC of their underlying resources.  Even though
+    // these aren't used outside of the method that creates them, they still must have hard
+    // references maintained to them.
     private SurfaceView mDummySurfaceView;
     private SurfaceTexture mDummySurfaceTexture;
 
+    /**
+     * Dedicated thread and associated runnable for calling into the detector with frames, as the
+     * frames become available from the camera.
+     */
     private Thread mProcessingThread;
     private FrameProcessingRunnable mFrameProcessor;
 
+    /**
+     * Map to convert between a byte array, received from the camera, and its associated byte
+     * buffer.  We use byte buffers internally because this is a more efficient way to call into
+     * native code later (avoids a potential copy).
+     */
     private Map<byte[], ByteBuffer> mBytesToByteBuffer = new HashMap<>();
 
+    //==============================================================================================
+    // Builder
+    //==============================================================================================
+
+    /**
+     * Builder for configuring and creating an associated camera source.
+     */
     public static class Builder {
         private final Detector<?> mDetector;
         private CameraSource mCameraSource = new CameraSource();
 
-
+        /**
+         * Creates a camera source builder with the supplied context and detector.  Camera preview
+         * images will be streamed to the associated detector upon starting the camera source.
+         */
         public Builder(Context context, Detector<?> detector) {
             if (context == null) {
                 throw new IllegalArgumentException("No context supplied.");
@@ -118,6 +187,10 @@ public class CameraSource {
             mCameraSource.mContext = context;
         }
 
+        /**
+         * Sets the requested frame rate in frames per second.  If the exact requested value is not
+         * not available, the best matching available value is selected.   Default: 30.
+         */
         public Builder setRequestedFps(float fps) {
             if (fps <= 0) {
                 throw new IllegalArgumentException("Invalid fps: " + fps);
@@ -136,7 +209,16 @@ public class CameraSource {
             return this;
         }
 
+        /**
+         * Sets the desired width and height of the camera frames in pixels.  If the exact desired
+         * values are not available options, the best matching available options are selected.
+         * Also, we try to select a preview size which corresponds to the aspect ratio of an
+         * associated full picture size, if applicable.  Default: 1024x768.
+         */
         public Builder setRequestedPreviewSize(int width, int height) {
+            // Restrict the requested range to something within the realm of possibility.  The
+            // choice of 1000000 is a bit arbitrary -- intended to be well beyond resolutions that
+            // devices can support.  We bound this to avoid int overflow in the code later.
             final int MAX = 1000000;
             if ((width <= 0) || (width > MAX) || (height <= 0) || (height > MAX)) {
                 throw new IllegalArgumentException("Invalid preview size: " + width + "x" + height);
@@ -146,6 +228,10 @@ public class CameraSource {
             return this;
         }
 
+        /**
+         * Sets the camera to use (either {@link #CAMERA_FACING_BACK} or
+         * {@link #CAMERA_FACING_FRONT}). Default: back facing.
+         */
         public Builder setFacing(int facing) {
             if ((facing != CAMERA_FACING_BACK) && (facing != CAMERA_FACING_FRONT)) {
                 throw new IllegalArgumentException("Invalid camera: " + facing);
@@ -154,28 +240,85 @@ public class CameraSource {
             return this;
         }
 
+        /**
+         * Creates an instance of the camera source.
+         */
         public CameraSource build() {
             mCameraSource.mFrameProcessor = mCameraSource.new FrameProcessingRunnable(mDetector);
             return mCameraSource;
         }
     }
 
+    //==============================================================================================
+    // Bridge Functionality for the Camera1 API
+    //==============================================================================================
+
+    /**
+     * Callback interface used to signal the moment of actual image capture.
+     */
     public interface ShutterCallback {
+        /**
+         * Called as near as possible to the moment when a photo is captured from the sensor. This
+         * is a good opportunity to play a shutter sound or give other feedback of camera operation.
+         * This may be some time after the photo was triggered, but some time before the actual data
+         * is available.
+         */
         void onShutter();
     }
 
+    /**
+     * Callback interface used to supply image data from a photo capture.
+     */
     public interface PictureCallback {
+        /**
+         * Called when image data is available after a picture is taken.  The format of the data
+         * is a jpeg binary.
+         */
         void onPictureTaken(byte[] data);
     }
 
+    /**
+     * Callback interface used to notify on completion of camera auto focus.
+     */
     public interface AutoFocusCallback {
+        /**
+         * Called when the camera auto focus completes.  If the camera
+         * does not support auto-focus and autoFocus is called,
+         * onAutoFocus will be called immediately with a fake value of
+         * <code>success</code> set to <code>true</code>.
+         * <p/>
+         * The auto-focus routine does not lock auto-exposure and auto-white
+         * balance after it completes.
+         *
+         * @param success true if focus was successful, false if otherwise
+         */
         void onAutoFocus(boolean success);
     }
 
+    /**
+     * Callback interface used to notify on auto focus start and stop.
+     * <p/>
+     * <p>This is only supported in continuous autofocus modes -- {@link
+     * Camera.Parameters#FOCUS_MODE_CONTINUOUS_VIDEO} and {@link
+     * Camera.Parameters#FOCUS_MODE_CONTINUOUS_PICTURE}. Applications can show
+     * autofocus animation based on this.</p>
+     */
     public interface AutoFocusMoveCallback {
+        /**
+         * Called when the camera auto focus starts or stops.
+         *
+         * @param start true if focus starts to move, false if focus stops to move
+         */
         void onAutoFocusMoving(boolean start);
     }
 
+    //==============================================================================================
+    // Public
+    //==============================================================================================
+
+    /**
+     * Stops the camera and releases the resources of the camera and underlying detector.
+     */
     public void release() {
         synchronized (mCameraLock) {
             stop();
@@ -183,6 +326,12 @@ public class CameraSource {
         }
     }
 
+    /**
+     * Opens the camera and starts sending preview frames to the underlying detector.  The preview
+     * frames are not displayed.
+     *
+     * @throws IOException if the camera's preview texture or display could not be initialized
+     */
     @RequiresPermission(Manifest.permission.CAMERA)
     public CameraSource start() throws IOException {
         synchronized (mCameraLock) {
@@ -192,7 +341,8 @@ public class CameraSource {
 
             mCamera = createCamera();
 
-
+            // SurfaceTexture was introduced in Honeycomb (11), so if we are running and
+            // old version of Android. fall back to use SurfaceView.
             if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB) {
                 mDummySurfaceTexture = new SurfaceTexture(DUMMY_TEXTURE_NAME);
                 mCamera.setPreviewTexture(mDummySurfaceTexture);
@@ -209,6 +359,13 @@ public class CameraSource {
         return this;
     }
 
+    /**
+     * Opens the camera and starts sending preview frames to the underlying detector.  The supplied
+     * surface holder is used for the preview so frames can be displayed to the user.
+     *
+     * @param surfaceHolder the surface holder to use for the preview frames
+     * @throws IOException if the supplied surface holder could not be used as the preview display
+     */
     @RequiresPermission(Manifest.permission.CAMERA)
     public CameraSource start(SurfaceHolder surfaceHolder) throws IOException {
         synchronized (mCameraLock) {
@@ -227,15 +384,26 @@ public class CameraSource {
         return this;
     }
 
+    /**
+     * Closes the camera and stops sending frames to the underlying frame detector.
+     * <p/>
+     * This camera source may be restarted again by calling {@link #start()} or
+     * {@link #start(SurfaceHolder)}.
+     * <p/>
+     * Call {@link #release()} instead to completely shut down this camera source and release the
+     * resources of the underlying detector.
+     */
     public void stop() {
         synchronized (mCameraLock) {
             mFrameProcessor.setActive(false);
             if (mProcessingThread != null) {
                 try {
-
+                    // Wait for the thread to complete to ensure that we can't have multiple threads
+                    // executing at the same time (i.e., which would happen if we called start too
+                    // quickly after stop).
                     mProcessingThread.join();
                 } catch (InterruptedException e) {
-                    PreyLogger.d("Frame processing thread interrupted on release.");
+                    Log.d(TAG, "Frame processing thread interrupted on release.");
                 }
                 mProcessingThread = null;
             }
@@ -247,6 +415,10 @@ public class CameraSource {
                 mCamera.stopPreview();
                 mCamera.setPreviewCallbackWithBuffer(null);
                 try {
+                    // We want to be compatible back to Gingerbread, but SurfaceTexture
+                    // wasn't introduced until Honeycomb.  Since the interface cannot use a SurfaceTexture, if the
+                    // developer wants to display a preview we must use a SurfaceHolder.  If the developer doesn't
+                    // want to display a preview we use a SurfaceTexture if we are running at least Honeycomb.
 
                     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB) {
                         mCamera.setPreviewTexture(null);
@@ -255,7 +427,7 @@ public class CameraSource {
                         mCamera.setPreviewDisplay(null);
                     }
                 } catch (Exception e) {
-                    PreyLogger.e("Failed to clear camera preview: " + e,e);
+                    Log.e(TAG, "Failed to clear camera preview: " + e);
                 }
                 mCamera.release();
                 mCamera = null;
@@ -263,10 +435,17 @@ public class CameraSource {
         }
     }
 
+    /**
+     * Returns the preview size that is currently in use by the underlying camera.
+     */
     public Size getPreviewSize() {
         return mPreviewSize;
     }
 
+    /**
+     * Returns the selected camera; one of {@link #CAMERA_FACING_BACK} or
+     * {@link #CAMERA_FACING_FRONT}.
+     */
     public int getCameraFacing() {
         return mFacing;
     }
@@ -280,7 +459,7 @@ public class CameraSource {
             int maxZoom;
             Camera.Parameters parameters = mCamera.getParameters();
             if (!parameters.isZoomSupported()) {
-                PreyLogger.d("Zoom is not supported on this device");
+                Log.w(TAG, "Zoom is not supported on this device");
                 return currentZoom;
             }
             maxZoom = parameters.getMaxZoom();
@@ -304,6 +483,15 @@ public class CameraSource {
         }
     }
 
+    /**
+     * Initiates taking a picture, which happens asynchronously.  The camera source should have been
+     * activated previously with {@link #start()} or {@link #start(SurfaceHolder)}.  The camera
+     * preview is suspended while the picture is being taken, but will resume once picture taking is
+     * done.
+     *
+     * @param shutter the callback for image capture moment, or null
+     * @param jpeg    the callback for JPEG image data, or null
+     */
     public void takePicture(ShutterCallback shutter, PictureCallback jpeg) {
         synchronized (mCameraLock) {
             if (mCamera != null) {
@@ -316,12 +504,33 @@ public class CameraSource {
         }
     }
 
+    /**
+     * Gets the current focus mode setting.
+     *
+     * @return current focus mode. This value is null if the camera is not yet created. Applications should call {@link
+     * #autoFocus(AutoFocusCallback)} to start the focus if focus
+     * mode is FOCUS_MODE_AUTO or FOCUS_MODE_MACRO.
+     * @see Camera.Parameters#FOCUS_MODE_AUTO
+     * @see Camera.Parameters#FOCUS_MODE_INFINITY
+     * @see Camera.Parameters#FOCUS_MODE_MACRO
+     * @see Camera.Parameters#FOCUS_MODE_FIXED
+     * @see Camera.Parameters#FOCUS_MODE_EDOF
+     * @see Camera.Parameters#FOCUS_MODE_CONTINUOUS_VIDEO
+     * @see Camera.Parameters#FOCUS_MODE_CONTINUOUS_PICTURE
+     */
     @Nullable
     @FocusMode
     public String getFocusMode() {
         return mFocusMode;
     }
 
+    /**
+     * Sets the focus mode.
+     *
+     * @param mode the focus mode
+     * @return {@code true} if the focus mode is set, {@code false} otherwise
+     * @see #getFocusMode()
+     */
     public boolean setFocusMode(@FocusMode String mode) {
         synchronized (mCameraLock) {
             if (mCamera != null && mode != null) {
@@ -338,12 +547,30 @@ public class CameraSource {
         }
     }
 
+    /**
+     * Gets the current flash mode setting.
+     *
+     * @return current flash mode. null if flash mode setting is not
+     * supported or the camera is not yet created.
+     * @see Camera.Parameters#FLASH_MODE_OFF
+     * @see Camera.Parameters#FLASH_MODE_AUTO
+     * @see Camera.Parameters#FLASH_MODE_ON
+     * @see Camera.Parameters#FLASH_MODE_RED_EYE
+     * @see Camera.Parameters#FLASH_MODE_TORCH
+     */
     @Nullable
     @FlashMode
     public String getFlashMode() {
         return mFlashMode;
     }
 
+    /**
+     * Sets the flash mode.
+     *
+     * @param mode flash mode.
+     * @return {@code true} if the flash mode is set, {@code false} otherwise
+     * @see #getFlashMode()
+     */
     public boolean setFlashMode(@FlashMode String mode) {
         synchronized (mCameraLock) {
             if (mCamera != null && mode != null) {
@@ -360,6 +587,24 @@ public class CameraSource {
         }
     }
 
+    /**
+     * Starts camera auto-focus and registers a callback function to run when
+     * the camera is focused.  This method is only valid when preview is active
+     * (between {@link #start()} or {@link #start(SurfaceHolder)} and before {@link #stop()} or {@link #release()}).
+     * <p/>
+     * <p>Callers should check
+     * {@link #getFocusMode()} to determine if
+     * this method should be called. If the camera does not support auto-focus,
+     * it is a no-op and {@link AutoFocusCallback#onAutoFocus(boolean)}
+     * callback will be called immediately.
+     * <p/>
+     * <p>If the current flash mode is not
+     * {@link Camera.Parameters#FLASH_MODE_OFF}, flash may be
+     * fired during auto-focus, depending on the driver and camera hardware.<p>
+     *
+     * @param cb the callback to run
+     * @see #cancelAutoFocus()
+     */
     public void autoFocus(@Nullable AutoFocusCallback cb) {
         synchronized (mCameraLock) {
             if (mCamera != null) {
@@ -373,6 +618,14 @@ public class CameraSource {
         }
     }
 
+    /**
+     * Cancels any auto-focus function in progress.
+     * Whether or not auto-focus is currently in progress,
+     * this function will return the focus position to the default.
+     * If the camera does not support auto-focus, this is a no-op.
+     *
+     * @see #autoFocus(AutoFocusCallback)
+     */
     public void cancelAutoFocus() {
         synchronized (mCameraLock) {
             if (mCamera != null) {
@@ -381,6 +634,12 @@ public class CameraSource {
         }
     }
 
+    /**
+     * Sets camera auto-focus move callback.
+     *
+     * @param cb the callback to run
+     * @return {@code true} if the operation is supported (i.e. from Jelly Bean), {@code false} otherwise
+     */
     @TargetApi(Build.VERSION_CODES.JELLY_BEAN)
     public boolean setAutoFocusMoveCallback(@Nullable AutoFocusMoveCallback cb) {
         if (Build.VERSION.SDK_INT < Build.VERSION_CODES.JELLY_BEAN) {
@@ -401,9 +660,19 @@ public class CameraSource {
         return true;
     }
 
+    //==============================================================================================
+    // Private
+    //==============================================================================================
+
+    /**
+     * Only allow creation via the builder class.
+     */
     private CameraSource() {
     }
 
+    /**
+     * Wraps the camera1 shutter callback so that the deprecated API isn't exposed.
+     */
     private class PictureStartCallback implements Camera.ShutterCallback {
         private ShutterCallback mDelegate;
 
@@ -415,7 +684,10 @@ public class CameraSource {
         }
     }
 
-
+    /**
+     * Wraps the final callback in the camera sequence, so that we can automatically turn the camera
+     * preview back on after the picture has been taken.
+     */
     private class PictureDoneCallback implements Camera.PictureCallback {
         private PictureCallback mDelegate;
 
@@ -432,6 +704,9 @@ public class CameraSource {
         }
     }
 
+    /**
+     * Wraps the camera1 auto focus callback so that the deprecated API isn't exposed.
+     */
     private class CameraAutoFocusCallback implements Camera.AutoFocusCallback {
         private AutoFocusCallback mDelegate;
 
@@ -443,6 +718,9 @@ public class CameraSource {
         }
     }
 
+    /**
+     * Wraps the camera1 auto focus move callback so that the deprecated API isn't exposed.
+     */
     @TargetApi(Build.VERSION_CODES.JELLY_BEAN)
     private class CameraAutoFocusMoveCallback implements Camera.AutoFocusMoveCallback {
         private AutoFocusMoveCallback mDelegate;
@@ -455,6 +733,11 @@ public class CameraSource {
         }
     }
 
+    /**
+     * Opens the camera and applies the user settings.
+     *
+     * @throws RuntimeException if the method fails
+     */
     @SuppressLint("InlinedApi")
     private Camera createCamera() {
         int requestedCameraId = getIdForRequestedCamera(mFacing);
@@ -494,7 +777,7 @@ public class CameraSource {
                     mFocusMode)) {
                 parameters.setFocusMode(mFocusMode);
             } else {
-                PreyLogger.i("Camera focus mode: " + mFocusMode + " is not supported on this device.");
+                Log.i(TAG, "Camera focus mode: " + mFocusMode + " is not supported on this device.");
             }
         }
 
@@ -506,7 +789,7 @@ public class CameraSource {
                     mFlashMode)) {
                 parameters.setFlashMode(mFlashMode);
             } else {
-                PreyLogger.i("Camera flash mode: " + mFlashMode + " is not supported on this device.");
+                Log.i(TAG, "Camera flash mode: " + mFlashMode + " is not supported on this device.");
             }
         }
 
@@ -515,7 +798,11 @@ public class CameraSource {
 
         camera.setParameters(parameters);
 
-
+        // Four frame buffers are needed for working with the camera:
+        //
+        //   one for the frame that is currently being executed upon in doing detection
+        //   one for the next pending frame to process immediately upon completing detection
+        //   two for the frames that the camera uses to populate future preview images
         camera.setPreviewCallbackWithBuffer(new CameraPreviewCallback());
         camera.addCallbackBuffer(createPreviewBuffer(mPreviewSize));
         camera.addCallbackBuffer(createPreviewBuffer(mPreviewSize));
@@ -525,6 +812,12 @@ public class CameraSource {
         return camera;
     }
 
+    /**
+     * Gets the id for the camera specified by the direction it is facing.  Returns -1 if no such
+     * camera was found.
+     *
+     * @param facing the desired camera (front-facing or rear-facing)
+     */
     private static int getIdForRequestedCamera(int facing) {
         CameraInfo cameraInfo = new CameraInfo();
         for (int i = 0; i < Camera.getNumberOfCameras(); ++i) {
@@ -536,6 +829,19 @@ public class CameraSource {
         return -1;
     }
 
+    /**
+     * Selects the most suitable preview and picture size, given the desired width and height.
+     * <p/>
+     * Even though we may only need the preview size, it's necessary to find both the preview
+     * size and the picture size of the camera together, because these need to have the same aspect
+     * ratio.  On some hardware, if you would only set the preview size, you will get a distorted
+     * image.
+     *
+     * @param camera        the camera to select a preview size from
+     * @param desiredWidth  the desired width of the camera preview frames
+     * @param desiredHeight the desired height of the camera preview frames
+     * @return the selected preview and picture size pair
+     */
     private static SizePair selectSizePair(Camera camera, int desiredWidth, int desiredHeight) {
         List<SizePair> validPreviewSizes = generateValidPreviewSizeList(camera);
 
@@ -568,8 +874,8 @@ public class CameraSource {
         private Size mPreview;
         private Size mPicture;
 
-        public SizePair(android.hardware.Camera.Size previewSize,
-                        android.hardware.Camera.Size pictureSize) {
+        public SizePair(Camera.Size previewSize,
+                        Camera.Size pictureSize) {
             mPreview = new Size(previewSize.width, previewSize.height);
             if (pictureSize != null) {
                 mPicture = new Size(pictureSize.width, pictureSize.height);
@@ -597,18 +903,18 @@ public class CameraSource {
      */
     private static List<SizePair> generateValidPreviewSizeList(Camera camera) {
         Camera.Parameters parameters = camera.getParameters();
-        List<android.hardware.Camera.Size> supportedPreviewSizes =
+        List<Camera.Size> supportedPreviewSizes =
                 parameters.getSupportedPreviewSizes();
-        List<android.hardware.Camera.Size> supportedPictureSizes =
+        List<Camera.Size> supportedPictureSizes =
                 parameters.getSupportedPictureSizes();
         List<SizePair> validPreviewSizes = new ArrayList<>();
-        for (android.hardware.Camera.Size previewSize : supportedPreviewSizes) {
+        for (Camera.Size previewSize : supportedPreviewSizes) {
             float previewAspectRatio = (float) previewSize.width / (float) previewSize.height;
 
             // By looping through the picture sizes in order, we favor the higher resolutions.
             // We choose the highest resolution in order to support taking the full resolution
             // picture later.
-            for (android.hardware.Camera.Size pictureSize : supportedPictureSizes) {
+            for (Camera.Size pictureSize : supportedPictureSizes) {
                 float pictureAspectRatio = (float) pictureSize.width / (float) pictureSize.height;
                 if (Math.abs(previewAspectRatio - pictureAspectRatio) < ASPECT_RATIO_TOLERANCE) {
                     validPreviewSizes.add(new SizePair(previewSize, pictureSize));
@@ -621,8 +927,8 @@ public class CameraSource {
         // of the preview sizes and hope that the camera can handle it.  Probably unlikely, but we
         // still account for it.
         if (validPreviewSizes.size() == 0) {
-            PreyLogger.d("No preview sizes have a corresponding same-aspect-ratio picture size");
-            for (android.hardware.Camera.Size previewSize : supportedPreviewSizes) {
+            Log.w(TAG, "No preview sizes have a corresponding same-aspect-ratio picture size");
+            for (Camera.Size previewSize : supportedPreviewSizes) {
                 // The null picture size will let us know that we shouldn't set a picture size.
                 validPreviewSizes.add(new SizePair(previewSize, null));
             }
@@ -690,7 +996,7 @@ public class CameraSource {
                 degrees = 270;
                 break;
             default:
-                PreyLogger.d("Bad rotation value: " + rotation);
+                Log.e(TAG, "Bad rotation value: " + rotation);
         }
 
         CameraInfo cameraInfo = new CameraInfo();
@@ -698,7 +1004,7 @@ public class CameraSource {
 
         int angle;
         int displayAngle;
-        if (cameraInfo.facing == Camera.CameraInfo.CAMERA_FACING_FRONT) {
+        if (cameraInfo.facing == CameraInfo.CAMERA_FACING_FRONT) {
             angle = (cameraInfo.orientation + degrees) % 360;
             displayAngle = (360 - angle); // compensate for it being mirrored
         } else {  // back-facing
@@ -713,11 +1019,23 @@ public class CameraSource {
         parameters.setRotation(angle);
     }
 
+    /**
+     * Creates one buffer for the camera preview callback.  The size of the buffer is based off of
+     * the camera preview size and the format of the camera image.
+     *
+     * @return a new preview buffer of the appropriate size for the current camera settings
+     */
     private byte[] createPreviewBuffer(Size previewSize) {
         int bitsPerPixel = ImageFormat.getBitsPerPixel(ImageFormat.NV21);
         long sizeInBits = previewSize.getHeight() * previewSize.getWidth() * bitsPerPixel;
         int bufferSize = (int) Math.ceil(sizeInBits / 8.0d) + 1;
 
+        //
+        // NOTICE: This code only works when using play services v. 8.1 or higher.
+        //
+
+        // Creating the byte array this way and wrapping it, as opposed to using .allocate(),
+        // should guarantee that there will be an array to work with.
         byte[] byteArray = new byte[bufferSize];
         ByteBuffer buffer = ByteBuffer.wrap(byteArray);
         if (!buffer.hasArray() || (buffer.array() != byteArray)) {
@@ -730,7 +1048,13 @@ public class CameraSource {
         return byteArray;
     }
 
+    //==============================================================================================
+    // Frame processing
+    //==============================================================================================
 
+    /**
+     * Called when the camera has a new preview frame.
+     */
     private class CameraPreviewCallback implements Camera.PreviewCallback {
         @Override
         public void onPreviewFrame(byte[] data, Camera camera) {
@@ -738,6 +1062,16 @@ public class CameraSource {
         }
     }
 
+    /**
+     * This runnable controls access to the underlying receiver, calling it to process frames when
+     * available from the camera.  This is designed to run detection on frames as fast as possible
+     * (i.e., without unnecessary context switching or waiting on the next frame).
+     * <p/>
+     * While detection is running on a frame, new frames may be received from the camera.  As these
+     * frames come in, the most recent frame is held onto as pending.  As soon as detection and its
+     * associated processing are done for the previous frame, detection on the mostly recently
+     * received frame will immediately start on the same thread.
+     */
     private class FrameProcessingRunnable implements Runnable {
         private Detector<?> mDetector;
         private long mStartTimeMillis = SystemClock.elapsedRealtime();
@@ -776,6 +1110,11 @@ public class CameraSource {
             }
         }
 
+        /**
+         * Sets the frame data received from the camera.  This adds the previous unused frame buffer
+         * (if present) back to the camera, and keeps a pending reference to the frame data for
+         * future use.
+         */
         void setNextFrame(byte[] data, Camera camera) {
             synchronized (mLock) {
                 if (mPendingFrameData != null) {
@@ -784,9 +1123,9 @@ public class CameraSource {
                 }
 
                 if (!mBytesToByteBuffer.containsKey(data)) {
-                    PreyLogger.d(
-                            "Skipping frame.  Could not find ByteBuffer associated with the image " +
-                                    "data from the camera.");
+                    Log.d(TAG,
+                        "Skipping frame.  Could not find ByteBuffer associated with the image " +
+                        "data from the camera.");
                     return;
                 }
 
@@ -801,6 +1140,20 @@ public class CameraSource {
             }
         }
 
+        /**
+         * As long as the processing thread is active, this executes detection on frames
+         * continuously.  The next pending frame is either immediately available or hasn't been
+         * received yet.  Once it is available, we transfer the frame info to local variables and
+         * run detection on that frame.  It immediately loops back for the next frame without
+         * pausing.
+         * <p/>
+         * If detection takes longer than the time in between new frames from the camera, this will
+         * mean that this loop will run without ever waiting on a frame, avoiding any context
+         * switching or frame acquisition time latency.
+         * <p/>
+         * If you find that this is using more CPU than you'd like, you should probably decrease the
+         * FPS setting above to allow for some idle time in between frames.
+         */
         @Override
         public void run() {
             Frame outputFrame;
@@ -814,7 +1167,7 @@ public class CameraSource {
                             // don't have it yet.
                             mLock.wait();
                         } catch (InterruptedException e) {
-                            PreyLogger.e("Frame processing loop terminated.", e);
+                            Log.d(TAG, "Frame processing loop terminated.", e);
                             return;
                         }
                     }
@@ -849,7 +1202,7 @@ public class CameraSource {
                 try {
                     mDetector.receiveFrame(outputFrame);
                 } catch (Throwable t) {
-                    PreyLogger.e("Exception thrown from receiver.", t);
+                    Log.e(TAG, "Exception thrown from receiver.", t);
                 } finally {
                     mCamera.addCallbackBuffer(data.array());
                 }
