+++ app/src/main/res/layout/activity_barcode.xml
@@ -5,7 +5,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    >
+    tools:context="com.prey.barcodereader.BarcodeActivity">
 
     <TextView
         android:layout_width="match_parent"
