+++ app/src/main/res/layout/barcode_capture.xml
@@ -8,16 +8,16 @@
     android:layout_height="match_parent"
     android:keepScreenOn="true">
 
-    <com.prey.barcode.ui.camera.CameraSourcePreview
-        android:id="@+id/preview"
+  <com.prey.barcodereader.ui.camera.CameraSourcePreview
+      android:id="@+id/preview"
+      android:layout_width="match_parent"
+      android:layout_height="match_parent">
+
+    <com.prey.barcodereader.ui.camera.GraphicOverlay
+        android:id="@+id/graphicOverlay"
         android:layout_width="match_parent"
-        android:layout_height="match_parent">
+        android:layout_height="match_parent" />
 
-        <com.prey.barcode.ui.camera.GraphicOverlay
-            android:id="@+id/graphicOverlay"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent" />
-
-    </com.prey.barcode.ui.camera.CameraSourcePreview>
+  </com.prey.barcodereader.ui.camera.CameraSourcePreview>
 
 </LinearLayout>
