+++ app/src/main/res/values/strings.xml
@@ -255,6 +255,6 @@
     <string name="barcode_success">Barcode read successfully</string>
     <string name="barcode_failure">No barcode captured</string>
     <string name="barcode_error">"Error reading barcode: %1$s"</string>
-
+    <string name="barcode_find_your_QR">Go to panel.preyproject.com/qr to find your QR code</string>
 
 </resources>
